LayoutTests/media/session/transient-interruptions-expected.txt

'Transient' interruption events should duck audio for 'Content' media sessions.


Waiting for Content media to begin playing.
EVENT(canplaythrough)
Media began playing.
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
EXPECTED (internals.mediaElementPlayerVolume(document.getElementById("video")) == '1') OK
Sending 'Transient' start-of-interruption notification.
RUN(internals.sendMediaSessionStartOfInterruptionNotification("transient"))
EXPECTED (internals.mediaSessionCurrentState(session) == 'interrupted') OK
EXPECTED (internals.mediaElementPlayerVolume(document.getElementById("video")) == '0.25') OK
Sending 'Transient' end-of-interruption notification.
RUN(internals.sendMediaSessionEndOfInterruptionNotification("transient"))
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
EXPECTED (internals.mediaElementPlayerVolume(document.getElementById("video")) == '1') OK
END OF TEST
